+++ src/middlewared/middlewared/plugins/pool.py
@@ -1,12 +1,100 @@
 import asyncio
-import libzfs
+import logging
+from datetime import datetime
 import os
+import subprocess
 import sysctl
 
-from datetime import datetime
+import bsd
+import libzfs
 
-from middlewared.schema import accepts, Int
-from middlewared.service import filterable, item_method, private, CRUDService
+from middlewared.job import JobProgressBuffer
+from middlewared.schema import accepts, Int, Str
+from middlewared.service import filterable, item_method, job, private, CRUDService
+from middlewared.utils import Popen, run
+
+logger = logging.getLogger(__name__)
+
+
+async def is_mounted(middleware, path):
+    mounted = await middleware.threaded(bsd.getmntinfo)
+    return any(fs.dest == path for fs in mounted)
+
+
+async def mount(dev, path, mntopts=None, fstype=None):
+    mount_cmd = ['/sbin/mount']
+
+    if isinstance(dev, str):
+        dev = dev.encode('utf-8')
+
+    if isinstance(path, str):
+        path = path.encode('utf-8')
+
+    if mntopts:
+        opts = ['-o', mntopts]
+    else:
+        opts = []
+
+    if fstype == 'ntfs':
+        mount_cmd = ['/usr/local/bin/ntfs-3g']
+        fstype = []
+    else:
+        fstype = ['-t', fstype] if fstype else []
+
+    proc = await Popen(
+        mount_cmd + opts + fstype + [dev, path],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        encoding='utf8',
+    )
+    output = await proc.communicate()
+
+    if proc.returncode != 0:
+        logger.debug("Mount failed (%s): %s", proc.returncode, output)
+        raise ValueError("Mount failed {0} -> {1}, {2}" .format(
+            proc.returncode,
+            output[0],
+            output[1]
+        ))
+    else:
+        return True
+
+
+class KernelModuleContextManager:
+    def __init__(self, module):
+        self.module = module
+
+    async def __aenter__(self):
+        if self.module is not None:
+            if not await self.module_loaded():
+                await run('kldload', self.module, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+                if not await self.module_loaded():
+                    raise Exception('Kernel module %r failed to load', self.module)
+
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
+        if self.module is not None:
+            try:
+                await run('kldunload', self.module, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+            except Exception:
+                pass
+
+    async def module_loaded(self):
+        return (await run('kldstat', '-n', self.module, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, check=False)).returncode == 0
+
+
+class MountFsContextManager:
+    def __init__(self, middleware, dev, path, *args):
+        self.middleware = middleware
+        self.dev = dev
+        self.path = path
+        self.args = args
+
+    async def __aenter__(self):
+        await mount(self.dev, self.path, *self.args)
+
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
+        if await is_mounted(self.middleware, self.path):
+            await self.middleware.threaded(bsd.unmount, self.path)
 
 
 class PoolService(CRUDService):
@@ -108,6 +196,91 @@ class PoolService(CRUDService):
         sysctl.filter('vfs.zfs.resilver_delay')[0].value = resilver_delay
         sysctl.filter('vfs.zfs.resilver_min_time_ms')[0].value = resilver_min_time_ms
         sysctl.filter('vfs.zfs.scan_idle')[0].value = scan_idle
+
+    @accepts(Str('volume'), Str('fs_type'), Str('dst_path'))
+    @job(lock=lambda args: 'volume_import')
+    async def import_disk(self, job, volume, fs_type, dst_path):
+        job.set_progress(None, description="Mounting")
+
+        src = os.path.join('/var/run/importcopy/tmpdir', os.path.relpath(volume, '/'))
+
+        if os.path.exists(src):
+            os.rmdir(src)
+
+        try:
+            os.makedirs(src)
+
+            async with KernelModuleContextManager({"ntfs": "fuse"}.get(fs_type)):
+                async with MountFsContextManager(self.middleware, volume, src, 'ro', fs_type):
+                    job.set_progress(None, description="Importing")
+
+                    line = [
+                        '/usr/local/bin/rsync',
+                        '--info=progress2',
+                        '--modify-window=1',
+                        '-rltvh',
+                        '--no-perms',
+                        src + '/',
+                        dst_path
+                    ]
+                    rsync_proc = await Popen(
+                        line, stdout=subprocess.PIPE, bufsize=0, preexec_fn=os.setsid,
+                    )
+                    stdout = b""
+                    try:
+                        progress_buffer = JobProgressBuffer(job)
+                        while True:
+                            line = await rsync_proc.stdout.readline()
+                            if line:
+                                stdout += line
+                                try:
+                                    proc_output = line.decode("utf-8", "ignore").strip()
+                                    prog_out = proc_output.split(' ')
+                                    progress = [x for x in prog_out if '%' in x]
+                                    if len(progress):
+                                        progress_buffer.set_progress(int(progress[0][:-1]))
+                                    elif not proc_output.endswith('/'):
+                                        if (
+                                            proc_output not in ['sending incremental file list'] and
+                                            'xfr#' not in proc_output
+                                        ):
+                                            progress_buffer.set_progress(None, extra=proc_output)
+                                except Exception:
+                                    logger.warning('Parsing error in rsync task', exc_info=True)
+                            else:
+                                break
+
+                        progress_buffer.flush()
+                        await rsync_proc.wait()
+                        if rsync_proc.returncode != 0:
+                            raise Exception("rsync failed with exit code %r" % rsync_proc.returncode)
+                    except asyncio.CancelledError:
+                        rsync_proc.kill()
+                        raise
+
+                    job.set_progress(100, description="Done", extra="")
+                    return stdout.decode("utf-8", "ignore")
+        finally:
+            os.rmdir(src)
+
+    """
+    These methods are hacks for old UI which supports only one volume import at a time
+    """
+
+    dismissed_import_disk_jobs = set()
+
+    @private
+    async def get_current_import_disk_job(self):
+        import_jobs = await self.middleware.call('core.get_jobs', [('method', '=', 'pool.import_disk')])
+        not_dismissed_import_jobs = [job for job in import_jobs if job["id"] not in self.dismissed_import_disk_jobs]
+        if not_dismissed_import_jobs:
+            return not_dismissed_import_jobs[0]
+
+    @private
+    async def dismiss_current_import_disk_job(self):
+        current_import_job = await self.get_current_import_disk_job()
+        if current_import_job:
+            self.dismissed_import_disk_jobs.add(current_import_job["id"])
 
 
 def setup(middleware):
